Lobby
A lobby was added, users can now choose to play/mute music, sound effects and show their fbs, they can also choose a track of their liking. They can now choose a username before entering the game, a button was added to join the game and a button was added to exit. More to come